engine: extract esm/minor_planets.js
